LayoutTests/ietestcenter/Javascript/15.2.3.11-4-12-expected.txt

15.2.3.11-4-12 (TestCases/chapter15/15.2/15.2.3/15.2.3.11/15.2.3.11-4-12.js) Object.isSealed returns false for all built-in objects (Number)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL ES5Harness.preconditionPassed should be true (of type boolean). Was undefined (of type undefined).
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
